creo pedirInfo.py agrego a read ahi
